test: split integration tests into separate files
